requirements.txt
paho-mqtt==2.1.0
setuptools==74.1.0